requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
atlassian-python-api==3.38.0
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dataclasses-json==0.5.7
Deprecated==1.2.14
Flask==2.3.2
Flask-SocketIO==5.3.4
frozenlist==1.3.3
fsspec==2023.6.0
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-aiplatform==1.26.0
google-cloud-bigquery==3.11.0
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.1
google-cloud-storage==2.9.0
google-cloud-texttospeech==2.14.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
html2text==2020.1.16
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.195
langchainplus-sdk==0.0.8
llama-hub==0.0.3
llama-index==0.6.22
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
python-dateutil==2.8.2
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
SQLAlchemy==2.0.15
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
vonage==3.5.2
vonage-jwt==1.0.0
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
